// src/app/(protected)/(common)/member/myhealth/healthProgramsResources/myHealthPrograms/models/health_program_type.ts
export enum HealthProgramType {
  TeladocSecondOption = 'TeladocSecondOption',
  SilverFit = 'SilverFit',
  TeladocBP = 'TeladocBP',
  HingeHealth = 'HingeHealth',
  TeladocHealthDiabetesManagement = 'TeladocHealthDiabetesManagement',
  QuestSelect = 'QuestSelect',
  TeladocMentalHealth = 'TeladocMentalHealth',
  TeladocPrimaryCareProvider = 'TeladocPrimaryCareProvider ',
  TalkToNurse = 'TalkToNurse',
  TeladocHealthGeneralUrgentCare = 'TeladocHealthGeneralUrgentCare',
  TeladocHealthDiabetesPrevention = 'TeladocHealthDiabetesPrevention',
  AbleTo = 'AbleTo',
}